WIP TS: tryingto isolate the client logic
